Update MEV capture hook artifacts.
