+++ v3/tasks/20250212-v3-mev-capture-hook/index.ts
@@ -1,0 +1,10 @@
+import { MevCaptureHookDeployment } from './input';
+import { Task, TaskRunOptions } from '@src';
+
+/* eslint-disable @typescript-eslint/no-non-null-assertion */
+export default async (task: Task, { force, from }: TaskRunOptions = {}): Promise<void> => {
+  const input = task.input() as MevCaptureHookDeployment;
+
+  const args = [input.Vault, input.BalancerContractRegistry];
+  await task.deployAndVerify('MevCaptureHook', args, from, force);
+};
